fix(home): Round 1 指摘 2 件を反映
Codex Round 1 の指摘:

1. home-screen.tsx の progressbar で aria-valuenow に生の attemptCount を
   渡していたため、目標超過日に aria-valuenow > aria-valuemax になる ARIA
   仕様違反。aria-valuenow を [0, dailyGoal] に clamp し、実数は
   aria-valuetext で補う。
2. JST 00:00 境界計算を daily-progress.ts と insights/history.ts で重複
   定義していた。src/lib/jst.ts に集約して両方から参照するように
   リファクタ (test も lib/jst.test.ts へ集約)。
