+++ src/server/home/daily-progress.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { fetchDailyProgress, jstStartOfToday } from "./daily-progress";
+import { fetchDailyProgress } from "./daily-progress";
 
 const queue: Array<() => unknown> = [];
 vi.mock("@/db/client", () => {
@@ -21,18 +21,6 @@ vi.mock("@/db/client", () => {
 
 beforeEach(() => {
   queue.length = 0;
-});
-
-describe("jstStartOfToday", () => {
-  it("UTC の 14:59 (= JST 23:59) は同日 00:00 JST を指す", () => {
-    const now = new Date("2026-04-19T14:59:00Z");
-    expect(jstStartOfToday(now).toISOString()).toBe("2026-04-18T15:00:00.000Z");
-  });
-
-  it("UTC の 15:00 (= JST 00:00) は翌日 00:00 JST を指す", () => {
-    const now = new Date("2026-04-19T15:00:00Z");
-    expect(jstStartOfToday(now).toISOString()).toBe("2026-04-19T15:00:00.000Z");
-  });
 });
 
 describe("fetchDailyProgress", () => {
